requirements.txt
langchain 
langchain-openai 
streamlit 
python-dotenv